simple-apps/currency_converter_bot/requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-tavily
langchain-community
langgraph
streamlit
python-dotenv
duckduckgo-search
six
urllib3
typing-extensions
